Refactor Ascend GDN attention metadata builder
