+++ vllm_ascend/patch/worker/__init__.py
@@ -36,7 +36,6 @@ import vllm_ascend.patch.worker.patch_qwen3_next_mtp  # noqa
 
 if not is_310p():
     import vllm_ascend.patch.worker.patch_qwen3_5  # noqa
-    import vllm_ascend.patch.worker.patch_gdn_attn  # noqa
     import vllm_ascend.patch.worker.patch_qwen3_dflash  # noqa
 
 import vllm_ascend.patch.worker.patch_rejection_sampler  # noqa
